New `ServerPayload`
This class will be used as the payload which is sent to the endpoint when called on. It can be subclasses and customised and works like a read-only dictionary.

To use custom the payload, refer this example:

```python

from discord.ext.ipc import ServerPayload

class CustomPayload(ServerPayload):
    def __init__(self, data):
        super().__init__(data)
        self.foo = data["bar"]
        # etc..

```

In the route function

```python

@route
async def get_remainder(self, payload: TimerPayload):
    # do some stuff here
```